keyswitch-tester-gui/requirements.txt
PySide6>=6.4.0
pyserial>=3.5 